feat(ci): add linting, CI pipeline, and binary release workflow
Fix all mypy strict errors (no-any-return, unused-ignore) across
api, auth, config, and models modules. Add __main__ guard for
PyInstaller compatibility.

- CI workflow: ruff check + format + mypy + pytest (Python 3.12/3.13)
- Release workflow: PyInstaller binaries (linux/macOS × x86_64/arm64)
  + PyPI publish via trusted publishers on v* tags
- curl|sh installer script for binary distribution
- Rename package to withings-health-cli (withings-cli taken on PyPI)

+++ src/withings_cli/models.py
@@ -92,7 +92,7 @@ class Measure(BaseModel):
 
     @property
     def real_value(self) -> float:
-        return self.value * (10**self.unit)
+        return float(self.value * (10**self.unit))
 
 
 class MeasureGroup(BaseModel):
